Add Pipeline feature
Pipeline provides delayed function calls. The so called
pipeline of actions is built using the builder pattern.
There are 3 actions:
- VoidFunctorAction; a function call without arguments
- TFunctorAction; a function call with a generic T argument
- DelayAction; a delay of x seconds

The Pipeline allows appending and prepending actions.
Prepended actions are added first in the chain while
appended actions are added last. The pipeline processes
the actions from first to last.