Add new Trashcan and Pass-Through model
Added new Trashcan and Pass-Through cabinet models!
Tried out a new way to import models from blender to unity but it didn't work that well.
Trashcan was made with glb format and Pass-Through with regular fbx.
Updated DoorHandle with animate option (Toggleable if we want the handle to move or not).